Fix memory leak during network race.
Adam caught a memory leak happening slowly over time.

Anytime a worker closes it's connection while a job is running there is
a reference loop between the Server::Client and the Server::Job object
in the ->{doing} element of the client. The fix is to clear the doing
list on disconnect of any kind.

This situation could be hit by the .5 second job pickup timeout in the
worker code, or by any other network latency or drop problems.

git-svn-id: http://code.sixapart.com/svn/gearman/trunk@370 011c6a6d-750f-0410-a5f6-93fdcd050bc4